+++ app/src/main/java/com/ynov/kotlin/rickmorty/data/entity/remote/Info.kt
@@ -1,8 +1,8 @@
 package com.ynov.kotlin.rickmorty.data.entity.remote
 
 data class Info (
-    val count: Long,
-    val pages: Long,
+    val count: Int,
+    val pages: Int,
     val next: String,
     val prev: String
 )
